[Refactor] to change the function createCard into the utils file
