+++ src/js/utils.js
@@ -10,3 +10,14 @@ export function getDataInJson(url) {
     .then(response => response.json())
     .catch(err => console.log(err));
 }
+
+export function createMealCard(meal) {
+  const templateCard = document.querySelector('.meal').content.querySelector('.meals-card').cloneNode(true);
+  const { strMeal, strMealThumb } = meal;
+
+  templateCard.querySelector('img').src = strMealThumb;
+  templateCard.querySelector('img').alt = strMeal;
+  templateCard.querySelector('p').textContent = strMeal;
+  console.log(templateCard);
+  return templateCard;
+}
